Use longs instead of ints to support larger key spaces.

+++ core/src/main/java/com/yahoo/ycsb/generator/CounterGenerator.java
@@ -17,31 +17,31 @@
 
 package com.yahoo.ycsb.generator;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Generates a sequence of integers 0, 1, ...
  */
 public class CounterGenerator extends NumberGenerator
 {
-	private final AtomicInteger counter;
+	private final AtomicLong counter;
 
 	/**
 	 * Create a counter that starts at countstart
 	 */
-	public CounterGenerator(int countstart)
+	public CounterGenerator(long countstart)
 	{
-		counter=new AtomicInteger(countstart);
+		counter=new AtomicLong(countstart);
 	}
 	
 	@Override
-  public Integer nextValue() 
+  public Long nextValue()
 	{
 		return counter.getAndIncrement();
 	}
 	
 	@Override
-	public Integer lastValue()
+	public Long lastValue()
 	{
 	    return counter.get() - 1;
 	}
